pyproject.toml
[project]
name = "cafe-order-system"
version = "0.1.0"
description = ""
authors = [
    { name = "Eric", email = "rozanov7475@yandex.by" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.32.2"
black = "^25.1.0"
pytest = "^8.3.4"
faker = "^36.1.1"
model-bakery = "^1.20.3"
django-debug-toolbar = "^5.0.1"

[tool.poe.env]
PYTHONPATH = "src"
APP_NAME = "cafe-order-system"
DC = "docker compose -p ${APP_NAME} -f docker-compose.dev.yml"
DC_PROD = "docker compose -p ${APP_NAME} -f docker-compose.prod.yml"

[tool.poe.tasks]
up = "$DC up --build"
down = "$DC down"
test = "pytest"
start = "gunicorn config.wsgi:application --bind 0.0.0.0:8080"
collectstatic = { shell = "cd src && python3 manage.py collectstatic --noinput" }
runserver = { shell = "cd src && python manage.py runserver" }
migrate = { shell = "cd src && python manage.py migrate" }
makemigrations = { shell = "cd src && python manage.py makemigrations" }
up_prod = "$DC_PROD up --build"
down_prod = "$DC_PROD down"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "tests.py test_*.py *_tests.py"
addopts = ["-rsxX", "-l", "--tb=short", "--strict-markers", "-s"]
markers = ["unit", "integration", "e2e"]
cache_dir = ".venv/cache/pytest"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
